requirements.txt
numpy==1.14.3
opencv-python==3.4.4.19
scipy==1.1.0
torch==1.0.0
easydict==1.9
torchvision==0.2.1
tqdm==4.28.1
Pillow==5.3.0
tensorboardX==1.5